+++ src/components/Card/MessageCard.jsx
@@ -11,7 +11,7 @@ import {
   DeleteContainer,
 } from "./MessageCardStyle";
 import DeleteButton from "../Buttons/DeleteButton";
-import { RelBadge } from "../Badge.jsx";
+import { RelBadge } from "../Badge/Badge";
 
 function MessageCard({
   recipientId,
